Implement new Minimal Design and Confirm request for deleting a job

+++ scripts/modules/ui.js
@@ -40,30 +40,26 @@ export default class UI {
           <h3 class="card__position-title">${job.position}</h3>
       </div>
     </div>
+
     <div class="card__right-wrapper">
       <input type="checkbox" class="supercheck" name="supercheck" ${UI.checkSuperjob(
         job.superjob
       )}>
-      <span class="tag">${job.location}</span>                        
-    </div> 
-    <div class="card__down-wrapper">
-      <a href="${job.link}" class="button--link" draggable="false" target="_blank">Link</a>
-      <button class="icon icon--arrow" draggable="false"> 
-        <img src="./assets/icons/down-arrow.png" alt="expand icon" draggable="false">
-      </button>
-      <div>
-        <button class="icon icon--edit" draggable="false">
-          <img src="./assets/icons/edit.png" alt="edit icon" draggable="false">
-        </button>
+      <div class="card__links-wrapper">
         <button class="icon icon--trash" draggable="false"> 
           <img src="./assets/icons/trash.svg" alt="trash icon" draggable="false">
         </button>
-      </div>
+        <button class="icon icon--link" draggable="false" onclick="location.href='${job.link}'" type="button">
+          <img src="./assets/icons/icon_link.svg" alt="link icon" draggable="false">
+        </button>
+      </div>                       
+    </div> 
     `;
 
     //Append card as the first child of the stage
     stage.insertAdjacentElement("afterbegin", card);
     UI.updateJobCounters();
+    UI.changeBorderSuperjob();
   }
 
   static deleteJobListener() {
@@ -73,9 +69,13 @@ export default class UI {
       icon.addEventListener("click", () => {
         const card = icon.closest(".card");
         const id = card.id;
-        card.remove();
-        Storage.deleteJob(id);
-        UI.updateJobCounters();
+        const result = confirm("Are you sure you want to delete this job?");
+        if (result) {
+          card.remove();
+          Storage.deleteJob(id);
+          UI.updateJobCounters();
+        }
+        
       })
     })
   };
@@ -86,6 +86,37 @@ export default class UI {
       const stage = counter.parentElement.parentElement;
       const stageContainer = stage.querySelector(".stage__card-container");
       counter.innerHTML = `${stageContainer.children.length} jobs`;
+    })
+  }
+
+  static changeBorderSuperjob () {
+    const cards = document.querySelectorAll(".card");
+    cards.forEach((card) => {
+      const stage = card.parentElement;
+      const stageCategory = stage.dataset.stage;
+      console.log(card, stageCategory);
+      if (card.classList.contains("card--supercard")) {
+        switch (stageCategory) {
+          case "wish-list":
+            card.style.border = "8px solid #F3DF26";
+            break;
+          case "applied":
+            card.style.border = "8px solid #4C6FFF";
+            break;
+          case "interview":
+            card.style.border = "8px solid #F34B26";
+            break;
+          case "offer":
+            card.style.border = "8px solid #26F36B";
+            break;
+          case "other":
+            card.style.border = "8px solid black";
+            break;
+        }
+      } else {
+        card.style.border = "8px solid #E6E6E6";
+      }
+      
     })
   }
